refactor: extract webview configuration and clients
