+++ app/src/main/java/com/fnostv/android4/MainActivity.java
@@ -1,15 +1,10 @@
 package com.fnostv.android4;
 
-import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.app.AlertDialog;
-import android.content.ActivityNotFoundException;
 import android.content.DialogInterface;
 import android.content.Intent;
-import android.graphics.Bitmap;
-import android.net.Uri;
 import android.net.http.SslError;
-import android.os.Build;
 import android.os.Bundle;
 import android.view.Gravity;
 import android.view.KeyEvent;
@@ -19,28 +14,29 @@ import android.view.Window;
 import android.view.WindowManager;
 import android.webkit.CookieManager;
 import android.webkit.CookieSyncManager;
-import android.webkit.DownloadListener;
 import android.webkit.SslErrorHandler;
-import android.webkit.WebChromeClient;
-import android.webkit.WebSettings;
 import android.webkit.WebView;
-import android.webkit.WebViewClient;
 import android.widget.FrameLayout;
 import android.widget.ProgressBar;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.fnostv.android4.config.ProfileStore;
 import com.fnostv.android4.config.ServerProfile;
 import com.fnostv.android4.util.Constants;
+import com.fnostv.android4.web.FnosChromeClient;
+import com.fnostv.android4.web.FnosDownloadListener;
+import com.fnostv.android4.web.FnosWebViewClient;
+import com.fnostv.android4.web.FullscreenVideoController;
+import com.fnostv.android4.web.LoginScript;
+import com.fnostv.android4.web.WebViewConfigurator;
+import com.fnostv.android4.web.WebViewEvents;
 
-public final class MainActivity extends Activity {
+public final class MainActivity extends Activity implements WebViewEvents {
     private FrameLayout root;
     private WebView webView;
     private ProgressBar progressBar;
     private TextView statusView;
-    private View customView;
-    private WebChromeClient.CustomViewCallback customViewCallback;
+    private FullscreenVideoController fullscreenVideoController;
     private ProfileStore store;
     private ServerProfile profile;
 
@@ -93,8 +89,8 @@ public final class MainActivity extends Activity {
             return true;
         }
         if (keyCode == KeyEvent.KEYCODE_BACK) {
-            if (customView != null) {
-                hideCustomView();
+            if (fullscreenVideoController.isShowing()) {
+                fullscreenVideoController.hide();
                 return true;
             }
             if (webView.canGoBack()) {
@@ -117,6 +113,7 @@ public final class MainActivity extends Activity {
         root.addView(webView, new FrameLayout.LayoutParams(
                 ViewGroup.LayoutParams.MATCH_PARENT,
                 ViewGroup.LayoutParams.MATCH_PARENT));
+        fullscreenVideoController = new FullscreenVideoController(root, webView);
 
         progressBar = new ProgressBar(this, null, android.R.attr.progressBarStyleHorizontal);
         FrameLayout.LayoutParams progressParams = new FrameLayout.LayoutParams(
@@ -138,42 +135,11 @@ public final class MainActivity extends Activity {
         setContentView(root);
     }
 
-    @SuppressLint("SetJavaScriptEnabled")
     private void configureWebView() {
-        webView.setFocusable(true);
-        webView.setFocusableInTouchMode(true);
-        webView.requestFocus();
-
-        WebSettings settings = webView.getSettings();
-        settings.setJavaScriptEnabled(true);
-        settings.setDomStorageEnabled(true);
-        settings.setDatabaseEnabled(true);
-        settings.setDatabasePath(getDir("webview-databases", MODE_PRIVATE).getPath());
-        settings.setLoadWithOverviewMode(true);
-        settings.setUseWideViewPort(true);
-        settings.setBuiltInZoomControls(false);
-        settings.setSupportZoom(false);
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR1) {
-            settings.setMediaPlaybackRequiresUserGesture(false);
-        }
-        settings.setUserAgentString(settings.getUserAgentString() + Constants.USER_AGENT_SUFFIX);
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
-            settings.setAllowFileAccessFromFileURLs(false);
-            settings.setAllowUniversalAccessFromFileURLs(false);
-        }
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.KITKAT) {
-            settings.setAppCacheEnabled(true);
-        }
-
-        CookieManager cookieManager = CookieManager.getInstance();
-        cookieManager.setAcceptCookie(true);
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-            cookieManager.setAcceptThirdPartyCookies(webView, true);
-        }
-
-        webView.setWebViewClient(new FnosWebViewClient());
-        webView.setWebChromeClient(new FnosChromeClient());
-        webView.setDownloadListener(new FnosDownloadListener());
+        WebViewConfigurator.configure(this, webView);
+        webView.setWebViewClient(new FnosWebViewClient(this, this));
+        webView.setWebChromeClient(new FnosChromeClient(this, fullscreenVideoController));
+        webView.setDownloadListener(new FnosDownloadListener(this));
     }
 
     private void loadProfileOrSettings() {
@@ -197,132 +163,59 @@ public final class MainActivity extends Activity {
     }
 
     private void enterFullScreen() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-            root.setSystemUiVisibility(
-                    View.SYSTEM_UI_FLAG_FULLSCREEN
-                            | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
-                            | View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY
-                            | View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
-                            | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
-                            | View.SYSTEM_UI_FLAG_LAYOUT_STABLE);
+        fullscreenVideoController.enterImmersiveMode();
+    }
+
+    @Override
+    public void onPageLoadStarted() {
+        progressBar.setVisibility(View.VISIBLE);
+    }
+
+    @Override
+    public void onPageLoadFinished(WebView view) {
+        progressBar.setVisibility(View.GONE);
+        CookieSyncManager.getInstance().sync();
+        if (profile.autoLogin && profile.username.length() > 0 && profile.password.length() > 0) {
+            view.loadUrl(LoginScript.build(profile));
         }
     }
 
-    private void hideCustomView() {
-        if (customView == null) {
+    @Override
+    public void onProgressChanged(int progress) {
+        progressBar.setProgress(progress);
+        progressBar.setVisibility(progress >= 100 ? View.GONE : View.VISIBLE);
+    }
+
+    @Override
+    public void onMainFrameError(String description) {
+        showStatus("加载失败：" + description + "\n按菜单键修改服务器设置");
+    }
+
+    @Override
+    public void onSslError(final SslErrorHandler handler, SslError error) {
+        if (profile != null && profile.trustSslErrors) {
+            handler.proceed();
             return;
         }
-        root.removeView(customView);
-        customView = null;
-        if (customViewCallback != null) {
-            customViewCallback.onCustomViewHidden();
-        }
-        customViewCallback = null;
-        webView.setVisibility(View.VISIBLE);
-        enterFullScreen();
+        new AlertDialog.Builder(MainActivity.this)
+                .setTitle("SSL 证书异常")
+                .setMessage("当前服务器证书无法验证，是否继续访问？")
+                .setPositiveButton("继续", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        handler.proceed();
+                    }
+                })
+                .setNegativeButton("取消", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        handler.cancel();
+                    }
+                })
+                .show();
     }
 
     private int dp(int value) {
         return (int) (value * getResources().getDisplayMetrics().density + 0.5f);
-    }
-
-    private final class FnosWebViewClient extends WebViewClient {
-        @Override
-        public boolean shouldOverrideUrlLoading(WebView view, String url) {
-            if (url.startsWith("http://") || url.startsWith("https://")) {
-                return false;
-            }
-            try {
-                startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse(url)));
-            } catch (ActivityNotFoundException ignored) {
-                Toast.makeText(MainActivity.this, "无法打开链接", Toast.LENGTH_SHORT).show();
-            }
-            return true;
-        }
-
-        @Override
-        public void onPageStarted(WebView view, String url, Bitmap favicon) {
-            progressBar.setVisibility(View.VISIBLE);
-            super.onPageStarted(view, url, favicon);
-        }
-
-        @Override
-        public void onPageFinished(WebView view, String url) {
-            progressBar.setVisibility(View.GONE);
-            CookieSyncManager.getInstance().sync();
-            if (profile.autoLogin && profile.username.length() > 0 && profile.password.length() > 0) {
-                view.loadUrl(LoginScript.build(profile));
-            }
-            super.onPageFinished(view, url);
-        }
-
-        @Override
-        public void onReceivedSslError(WebView view, final SslErrorHandler handler, SslError error) {
-            if (profile != null && profile.trustSslErrors) {
-                handler.proceed();
-                return;
-            }
-            new AlertDialog.Builder(MainActivity.this)
-                    .setTitle("SSL 证书异常")
-                    .setMessage("当前服务器证书无法验证，是否继续访问？")
-                    .setPositiveButton("继续", new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-                            handler.proceed();
-                        }
-                    })
-                    .setNegativeButton("取消", new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-                            handler.cancel();
-                        }
-                    })
-                    .show();
-        }
-
-        @Override
-        public void onReceivedError(WebView view, int errorCode, String description, String failingUrl) {
-            showStatus("加载失败：" + description + "\n按菜单键修改服务器设置");
-        }
-
-    }
-
-    private final class FnosChromeClient extends WebChromeClient {
-        @Override
-        public void onProgressChanged(WebView view, int newProgress) {
-            progressBar.setProgress(newProgress);
-            progressBar.setVisibility(newProgress >= 100 ? View.GONE : View.VISIBLE);
-        }
-
-        @Override
-        public void onShowCustomView(View view, CustomViewCallback callback) {
-            if (customView != null) {
-                callback.onCustomViewHidden();
-                return;
-            }
-            customView = view;
-            customViewCallback = callback;
-            webView.setVisibility(View.GONE);
-            root.addView(view, new FrameLayout.LayoutParams(
-                    ViewGroup.LayoutParams.MATCH_PARENT,
-                    ViewGroup.LayoutParams.MATCH_PARENT));
-            enterFullScreen();
-        }
-
-        @Override
-        public void onHideCustomView() {
-            hideCustomView();
-        }
-    }
-
-    private final class FnosDownloadListener implements DownloadListener {
-        @Override
-        public void onDownloadStart(String url, String userAgent, String contentDisposition, String mimetype, long contentLength) {
-            try {
-                startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse(url)));
-            } catch (ActivityNotFoundException ignored) {
-                Toast.makeText(MainActivity.this, "没有可用的下载应用", Toast.LENGTH_SHORT).show();
-            }
-        }
     }
 }
