+++ app/src/main/java/com/fnostv/android4/web/FnosWebViewClient.java
@@ -1,0 +1,57 @@
+package com.fnostv.android4.web;
+
+import android.content.ActivityNotFoundException;
+import android.content.Context;
+import android.content.Intent;
+import android.graphics.Bitmap;
+import android.net.Uri;
+import android.net.http.SslError;
+import android.webkit.SslErrorHandler;
+import android.webkit.WebView;
+import android.webkit.WebViewClient;
+import android.widget.Toast;
+
+public final class FnosWebViewClient extends WebViewClient {
+    private final Context context;
+    private final WebViewEvents events;
+
+    public FnosWebViewClient(Context context, WebViewEvents events) {
+        this.context = context;
+        this.events = events;
+    }
+
+    @Override
+    public boolean shouldOverrideUrlLoading(WebView view, String url) {
+        if (url.startsWith("http://") || url.startsWith("https://")) {
+            return false;
+        }
+        try {
+            context.startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse(url)));
+        } catch (ActivityNotFoundException ignored) {
+            Toast.makeText(context, "无法打开链接", Toast.LENGTH_SHORT).show();
+        }
+        return true;
+    }
+
+    @Override
+    public void onPageStarted(WebView view, String url, Bitmap favicon) {
+        events.onPageLoadStarted();
+        super.onPageStarted(view, url, favicon);
+    }
+
+    @Override
+    public void onPageFinished(WebView view, String url) {
+        events.onPageLoadFinished(view);
+        super.onPageFinished(view, url);
+    }
+
+    @Override
+    public void onReceivedSslError(WebView view, SslErrorHandler handler, SslError error) {
+        events.onSslError(handler, error);
+    }
+
+    @Override
+    public void onReceivedError(WebView view, int errorCode, String description, String failingUrl) {
+        events.onMainFrameError(description);
+    }
+}
