+++ app/src/main/java/com/fnostv/android4/web/FullscreenVideoController.java
@@ -1,0 +1,64 @@
+package com.fnostv.android4.web;
+
+import android.os.Build;
+import android.view.View;
+import android.view.ViewGroup;
+import android.webkit.WebChromeClient;
+import android.webkit.WebView;
+import android.widget.FrameLayout;
+
+public final class FullscreenVideoController {
+    private final FrameLayout root;
+    private final WebView webView;
+    private View customView;
+    private WebChromeClient.CustomViewCallback customViewCallback;
+
+    public FullscreenVideoController(FrameLayout root, WebView webView) {
+        this.root = root;
+        this.webView = webView;
+    }
+
+    public boolean isShowing() {
+        return customView != null;
+    }
+
+    public void show(View view, WebChromeClient.CustomViewCallback callback) {
+        if (customView != null) {
+            callback.onCustomViewHidden();
+            return;
+        }
+        customView = view;
+        customViewCallback = callback;
+        webView.setVisibility(View.GONE);
+        root.addView(view, new FrameLayout.LayoutParams(
+                ViewGroup.LayoutParams.MATCH_PARENT,
+                ViewGroup.LayoutParams.MATCH_PARENT));
+        enterImmersiveMode();
+    }
+
+    public void hide() {
+        if (customView == null) {
+            return;
+        }
+        root.removeView(customView);
+        customView = null;
+        if (customViewCallback != null) {
+            customViewCallback.onCustomViewHidden();
+        }
+        customViewCallback = null;
+        webView.setVisibility(View.VISIBLE);
+        enterImmersiveMode();
+    }
+
+    public void enterImmersiveMode() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
+            root.setSystemUiVisibility(
+                    View.SYSTEM_UI_FLAG_FULLSCREEN
+                            | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
+                            | View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY
+                            | View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
+                            | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
+                            | View.SYSTEM_UI_FLAG_LAYOUT_STABLE);
+        }
+    }
+}
